Refactored daac_data_subscriber.run_rtc_download() into the run_download() method of the AsfDaacRtcDownload class
This was done to bring the interface of AsfDaacRtcDownload in line with the other inheritors of DaacDownload, and simplify the logic within daac_data_subscriber.py
